refactoring spoilarr from monolith code structure
