+++ spoilarr/types.py
@@ -1,0 +1,3 @@
+from typing import Literal
+
+RequestType = Literal["discord_deleted_user", "owner", "user", "user_strict"]
